API_PRE_REGISTRO/requirements.txt
fastapi 
redis-om
uvicorn[standard]
# "fastapi[standard]"
requests 
pymongo
pydantic-settings 
python-dotenv
motor




